파일 정리 및 포스트서비스 변경
