--- .gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,0 @@
-#Tue Feb 18 10:33:49 KST 2025
-gradle.version=8.12.1
