--- src/main/java/com/newsgroup/newsfeed/entity/PostComments.java
@@ -1,25 +1,0 @@
-package com.newsgroup.newsfeed.entity;
-
-import com.newsgroup.newsfeed.config.BaseEntity;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor
-public class PostComments extends BaseEntity {
-
-    @Id
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "users_id", nullable = false)
-    private Users user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "posts_id", nullable = false)
-    private Posts posts;
-
-    private String content;
-}
